Added a method for `transform`.
This method is useful when you have a hierarchy of cell references
that is not strictly a tree, but may have branches that return to
the trunk. Practically speaking, it allows for easier placement
of Josephson junctions when you need to place them in cells that may
have been reflected, etc.